+++ src/components/Output.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import ClipboardCopy from '@/components/ClipboardCopy.vue'
+import Pre from './Pre.vue'
 
 defineProps({
   data: {
@@ -15,10 +16,7 @@ defineProps({
 
 <template>
   <div class="space-y-2">
-    <pre
-      class="text-s max-h-96 overflow-scroll border border-black bg-neutral-100 dark:border-white dark:bg-neutral-900"
-      >{{ data }}</pre
-    >
+    <Pre>{{ data }}</Pre>
     <ClipboardCopy v-if="enableCopy" :data="data" />
   </div>
 </template>